Warcraft API backend updated.
The Warcraft API had to have new credentials created due to Blizzard changing their API portal. These changes have been implemented and are working fine as far as I can tell. Closes #2.

The database.py file has been broken into multiple files to handle DB methods for the different tables/classes. This will allow for easier search and modification of code that is related to specific modules.

casper.py had a few typos fixed.
