+++ cogs_production/reminders/reminders_commands.py
@@ -7,7 +7,8 @@ import asyncio
 import datetime
 
 from discord.ext import commands
-from database.database import RemindersDatabaseMethods, CountdownEventsDatabaseMethods
+from database.reminders import RemindersDatabaseMethods
+from database.countdownevents import CountdownEventsDatabaseMethods
 
 
 class Reminder:
@@ -24,7 +25,7 @@ class Reminder:
         while not self.casper.is_closed():
             guilds = self.casper.guilds
             for guild in guilds:
-                reminders = await RemindersDatabaseMethods.get_reminders(guild.id)
+                reminders = await RemindersDatabaseMethods.auto_get_reminders(guild.id)
                 if len(reminders) != 0:
                     for reminder in reminders:
                         ch = self.casper.get_channel(reminder.channel)
@@ -76,6 +77,13 @@ class Reminder:
 
     @commands.command()
     async def addevent(self, ctx, event_date: str, *, event_name: str):
+        """
+        Add an event to be tracked by the countdown command.
+        :param ctx: Invoked context.
+        :param event_date: Date of the event.
+        :param event_name: Name of the event.
+        :return: On successful event creation, confirmation message. Else, usage message.
+        """
         usage_msg = ('Please format your date as MM/DD/YYYY.\n'
                      'Ex: casper addevent 01/01/2019 Happy New Year!')
         if '/' not in event_date:
@@ -96,7 +104,48 @@ class Reminder:
             return await ctx.send(usage_msg)
 
     @commands.command()
+    async def removeevent(self, ctx, event_date: str, *, event_name: str):
+        """
+        Removes an event from the countdown tracker.
+        :param ctx: Invoked context.
+        :param event_date: Date of the event.
+        :param event_name: Name of the event.
+        :return: On successful event deletion, confirmation message. Else, error message.
+        """
+        usage_msg = ('Please format your date as MM/DD/YYYY.\n'
+                     'Ex: casper removeevent 01/01/2019 Happy New Year!')
+        if '/' not in event_date:
+            return await ctx.send(usage_msg)
+        try:
+            month, day, year = event_date.split('/')
+            month = int(month)
+            day = int(day)
+            year = int(year)
+        except ValueError:
+            return await ctx.send(usage_msg)
+        if 0 < month < 13 and 0 < day < 32 and 0 < year < 100000:
+            event_date = datetime.date(year, month, day)
+            if await CountdownEventsDatabaseMethods.remove_event(
+                    ctx.guild.id, event_name, event_date):
+                return await ctx.send(f'Countdown for {event_name} on {event_date} '
+                                      f'removed.')
+            else:
+                return await ctx.send(f'Could not find event matching "{event_name}" on '
+                                      f'{event_date}. To delete an event, you must make '
+                                      f'sure the date and name match **__exactly__**.')
+        else:
+            return await ctx.send(usage_msg)
+
+    @commands.command()
     async def countdown(self, ctx, *, event_name: str=None):
+        """
+        Get the current countdown to a given event, or all events if no event name is
+        given.
+        :param ctx: Invoked context.
+        :param event_name: Name of the event to countdown to.
+        :return: If event_name is given, all events which match. Else, every event for
+            the particular discord server.
+        """
         if event_name:
             events = await CountdownEventsDatabaseMethods.get_event(
                 ctx.guild.id, event_name)
@@ -107,10 +156,10 @@ class Reminder:
             for event in events:
                 days_til_event = (event.event_date - datetime.date.today()).days
                 if days_til_event < 0:
-                    await CountdownEventsDatabaseMethods.remove_event(ctx.guild.id,
-                                                                      event.id)
+                    await CountdownEventsDatabaseMethods.auto_remove_event(ctx.guild.id,
+                                                                           event.id)
                     output_str += (f'The event "{event.event_name}" has passed and '
-                                   f'has been removed from the countdown tracker.')
+                                   f'has been removed from the countdown tracker.\n')
                     continue
                 output_str += (f'There are {days_til_event} days left until '
                                f'{event.event_name}.\n\n')
